Move queries from the repository to specific module. Rename queries.
